src/main.py: Check if we uploaded a multiple chunk file before deleting the resume file

+++ src/main.py
@@ -179,7 +179,9 @@ class transferHandler:
 
         finalData = self.tgObject[sFile].uploadFiles(fileData)
 
-        os.remove(os.path.join(self.data_path, "resume_{}".format(sFile)))
+        if len(finalData['fileData']['fileID']) > 1:
+            os.remove(os.path.join(self.data_path, "resume_{}".format(sFile)))
+
         with open(os.path.join(self.data_path, "index_{}".format(sFile)), 'w') as f:
             f.write(str(finalData['index']))
